Add visual studio source files
the windows/adolc.sln file is the solution file which can be
used to compile both sparse and nosparse versions.
windows/adolc.vcxproj is the adolc project file containing the
configuration for both sparse and nosparse builds, the sparse build
requires the ThirdParty/ColPack/ColPack.vcxproj project to be
compiled first.

In order to compile ColPack, just unpack the ColPack sources into
ThirdParty/ColPack and apply the patch
Thirparty/ColPack/colpack_vcxcompile.patch
which is needed for colpack to be compiled in VC++.
After this Visual Studio will be able to build the sparse version
of adolc which contains ColPack as a static library within.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.hu-berlin.de>