ColPack and ADOL-C memory management conflicts
Recall the latest changes we asked you to make in the ADOL-C functions
generate_seed_jac and generate_seed_hess: you had to copy the seed matrix
so that ADOL-C could later free the associated memory.
To avoid the copying there, and allow deallocation of the seed matrix by ADOL-C,
we have added in ColPack GenerateSeedJacobian and GenerateSeedHessian functions
whose memory deallocation is not managed by ColPack, but by the caller, in this case ADOL-C.
These functions have the suffix "_unmanaged" in their names.
In addition, we have modified the GenerateSeedJacobian and  GenerateSeedHessian routines
so that they would not clear and build the bipartite graph afresh, but rather rely on the graph built via
the constructors BipartiteGraphPartialColoringInterface(SRC_MEM_ADOLC, ...) and
GraphColoringInterface(SRC_MEM_ADOLC,...).

Signed-off-by: Duc Nguyen Chi <dnguy010@gmail.com>
Acked-by: Assefaw Gebremedhin <agebreme@purdue.edu>
Signed-off-by: Kshitij Kulshreshtha <kshitij@math.hu-berlin.de>
Cc: Andrea Walther <andrea.walther@uni-paderborn.de>