Allow repeated use of the same memory for sparse_hess and sparse_jac
This is in conjunction with the new ColPack functions from version 1.0.2
If the pointers rind, cind and values passed to sparse_hess and spase_jac
not NULL they will be assumed to be allocated properly by the user.
Any segmentation-violation due to incorrect allocation is the responsibility
of the user. If any one of rind, cind or values is NULL then the other
non-NULL pointers will be freed and all three will be reallocated

Older Colpack than 1.0.2 cannot be used with this.

Feature-request-by: Andrey Romanenko <andrey.romanenko@ciengis.com>
Cc: Assefaw Gebremedhin <agebreme@purdue.edu>
    Andrea Walther <andrea.walther@uni-paderborn.de>
Signed-off-by: Kshitij Kulshreshtha <kshitij@math.hu-berlin.de>