Handle the pattern only case in all processes
for sparse_jac with options[2] == -1 and for sparse_hess with
repeat == -1, only the pattern is computed and stored. The Jacobian
or the Hessian itself is not propagated. This should be handled in
all processes, otherwise the id != 0 processes will continue to try
to propogate derivatives and hang in a send/recv.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>