check for zero blocks separately
iter=indexFree.erase(iter) advances iter to the next element, then the
for loop advances it again. So we potentially miss a block, which could
either be a tail block or a zero block. Just stop as soon as we find tail
and we remove the zero blocks after allocating the new tail block, if any
are remaining.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>