Move some options out of usrparms.h
These should be set using the configure script arguments and are
now in a new generated header file that will be installed via
make install. This include ATRIG_ERF, and the typedefs for locint/revreal
A new option ADOLC_ADVANCED_BRANCHING is also created.

ADOLC_USE_CALLOC goes into config.h as this is only required for build
but not after installation

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>