Allow adouble objects to be initialized late
This functionality is specially required for large codes that allocate
memory using malloc/calloc and we only wish to substitute the double
datatype with adouble for tracing purposes. With malloc/calloc the
adouble constructor is not called, creating problems later in the code.
Late initialization defers some of the initialization to the first use
of an allocated space as an adouble object. This is only possible because
the class heirarchy in ADOL-C does not require polymorphism, and virtual
tables to be allocated, which can only be done in a chained superclass
constructor call.

The functionality is disabled by default and can be enabled by a configure
switch --enable-lateinit

This code is experimental and has only be tested with particular codes.

From: Max Sagebaum <sagebaum@mathcces.rwth-aachen.de>
Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>
CC: Andrea Walther <andrea.walther@uni-paderborn.de>
CC: Nicolas Gauger <gauger@mathcces.rwth-aachen.de>