reimplement adubp_from_adub using the isInit flag
I didn't want to use a const_cast<> but there seems to be no way around.
This implementation simply steals the location from the adub argument
and sets it as the location of the adub* result. The isInit=false flag tells
the ~adub destructor not to free the location.
This fixes the problems where multiple locations are allocated and used
for results, like in sin/cos etc.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>