create space and methods for tracing the parameters
we can use a separate StoreManager to manage the indices
of the parameters during tracing. Do this in GlobalTapeVars
just like for the active store.

The difference is that parameters don't get overwritten so
the management is really much simpler. This indices will all
only be freed once taping is complete.

Also add constructors and methods to tag parameters in user
code in a new file.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>