Store and retrieve parameters from disk tapes
At the end of tracing in stop_trace we can copy the
parameters from the tracing store in GlobalTapeVars
to the store in PersistentTapeInfos. This way, we can
trace a different function after this one, with a
different set of parameters.

Also store these parameters contiguously at the end of
the value tape for persistence. These should be read at
the same time as the stats are read from disk tapes
when opening them, so that we can call forward and
reverse directly without any changes.

One can call set_param_vec() on any tape, if one wishes
to replace the parameters with new ones, before calling
forward. Taylor stack is invalidated if one does this,
so forward must be called before reverse.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>