convert all ipopt examples to use the parameter interface
since lagrange multipliers in optimization only occur
as linear parameters we do not require any derivatives
with respect to these. However they change in each
iteration of the optimizer and need to be considered
when computing the hessian of the lagrangian function.
so the use of the parameter interface is an ideal
solution here.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>