Fix ext_diff_fct so that users don't do memory allocation
the memory depends on the size of input and output. so it
can change depending on how many of those there are.
make it dependent on max_m, max_n and do deallocation
and allocation within call_ext_fct

if the maximal size is allocated then smaller sizes will
work automatically.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>