add header file for the more generic external functions interface
This interface will exist parallel to the previous interface
and might even use some of the existing code.
The main reason for designing this is to lift the limitation
of one contigously allocated adouble input array and one
similar output array. Here we can use several arrays for input
and output.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>