Revert "Also use boost pool_allocator in tapeless mode"
As noticed when using ADIC runtime (which is basically the same usage)
pool_allocator does not help but actually harms the performance in this
case. This is because the size of the derivative array is not a compile
time constant and the pool allocator requires object sizes that it
manages to be known at compile time.

This reverts commit 3440570f80743bb67930745ba5a72b0f798fca8b.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>