add explicit downcasts from badouble to double
These are to be used only in very special cases.
Like when I/O is required, different from the default
I/O provided. Or to make certain variables passive.
The explicit specifier makes sure that the compiler
does not automatically do any type changes without
the user saying so in the code.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>