add functions to import and export functioning traces to ASCII files
this add function to write an existing trace to an ASCII file and
also to read it back from the ASCII file and reevaluate this new trace.

AdjoinableMPI functionality is not supported since AdjoinableMPI has
its own tape, which ADOL-C cannot export or import.

External differentiation and checkpointing may or may not work. This
has not been tested yet.

Signed-off-by: Kshitij Kulshreshtha <kshitij@math.upb.de>