pass explicit libdir to setup.py and fix out-of-tree build even more
libdir passed will only be used to find libadolc.so for linking,
so that users who have explicitly set this in ./configure can still
install.

But python will still install its own module in:
${prefix}/lib64/python${py_version}/site-packages
and ignore the given libdir for installation.

This is the default python distutils behaviour and even if I could
I will not change it.

python distutils and autoconf/automake are not really compatible with
each other.