Don't short-circuit multiplication
The code contained constructions like

  a = (b==0) ? 0 : b*c;

This is less readable than a simple

  a = b*c;

Supposedly, the longer construction is there in the hope of
faster code.  However, the internet claims that it does not
lead to speedup.  See, e.g.,

  https://stackoverflow.com/questions/25897141/is-multiplication-by-zero-a-special-case

and

  https://stackoverflow.com/questions/15214673/can-floating-point-multiplication-by-zero-be-optimised-at-runtime

This makes sense to me.  On some processors, checking for zeros before
multiplications may be a smart thing to do, but then the compiler knows
best about it and is able to insert the necessary checks itself.