Improve const correctness of drivers
This changes several input arguments for driver functions to `const`. The patch
only touches the functions that are documented in the freely available documentation
https://www3.math.tu-berlin.de/Vorlesungen/SS06/AlgoDiff/adolc-110.pdf.

The changed parameters are documented to be input paramters. Hence declaring
them const should be fine. In fact this required that some internal function
arguments also had to be made `const`.

This patch leaves out the functions that are ot documented in the linked
manual. It also keeps the `lagrange` argument of `vec_jac()` non-const.
Changing the latter would require to modify the 4th argument of the internal
`fos_reverse()` method. However, this argument is used for write access
and hence the `lagrange` parameter seems to be modified in fact.

*Disclaimer:*
While I tested all changes with `make test` (which does not lead to any compiler
warnings about `const`-ness issues), I have no idea about the Adol-C internals.
Hence all changes-especially to the internal methods `fos_forward()`, `fos_forward_nk()`,
and `hos_reverse()`-should be carefully reviewed by someone how understands the internals.