Move StoreManager definitions into a dedicated file
Formerly they were contained in tape_handling.cpp,
along with a lot of other stuff.  To make the code
easier to understand this commit takes everything
that has to do with StoreManager classes, and moves it
into a new file storemanager.cpp.  The corresponding
header with the declarations even existed previously.