Remove the clearCurrentTape method
What it did was zeroing out the ADOLC_CURRENT_DATE_INFOS
and ADOLC_CURRENT_TAPE_INFOS_FALLBACK objects.  However,
it did this in a rather convoluted way:  By creating a new
object (which is zeroed because its constructor calls
initTapeInfos) and then assigning that object.

However, it is much simpler to call initTapeInfos directly
for ADOLC_CURRENT_DATE_INFOS(_FALLBACK), and this is what
this patch does.  The clearCurrentTape method becomes
obsolete and is removed.