Really move the data when assigning adub to adouble
The adub data type is used to only store temporary values,
and once an adub object is assigned to an adouble object
it shall never be used again.  Therefore, assignment of
adub to adouble can be implemented just like move-assignment,
without having to touch the tape at all.

For this to work, the method signature has to change from
(const adub&) to (adub&), because the value that is assigned
is now modified (it is invalidated).