This commit introduces the ValueTape; rm OpenMp for the moment.
The ValueTape is in principle the extracted behavior of the previous taping mechanism that were spread over the whole codebase via macros.

The class ValueTape serves as the composition and interface of the globalvariables, tapeinfos and persistant tape infos, as well as for the buffers of external functions and checkpoints. In addition, there are several changes to improve the code quality but not its overall behavior. For example the enums are now part of the classes (struct) to which they logically belong. Here it should be noted, that most of the older initialization techniques has been replaced by proper constructors or destructors. Or classes are only moveable but not copyable, because there should never be a tape with a similar id that points to the same buffer etc.