Introducing a improved Errorhandler ADOLCError, replaces FatalError
ADOLCError is utilized together with ADOLC_ERRORS (enum) and fail to provide access to a unified error handling. Users should use the fail function together with an error code (ADOLC_ERRORS), the source_location (providing the file, line and function where the error occured, C++20 feature) and potential infos like tapeId given via FailInfo object. Inside fail, based on the ADOLC_ERRORS a message, infos and the source location creates an ADOLCError instance. The ADOLCError as a child of runtime_erorr is throw printing out all the infomation provided.

At the current time, the source_location::current() does not work properly when handing to fail as default argument. Thats why this method should be called and passed to fail manually.