Adapt to ValueTape, introduce new constructors, separat location handling to tape_location; rm OpenMp for the moment
The classes `adouble` and `pdouble` have now a `tape_location` that automatically creates a loction on the currentTape

Location handling is now extracted from `adouble` and `pdouble` to `tape_location` adhering
separation of concerns. All calls to `next_loc` and `free_loc` are done in `tape_location`.

`tape_location` is now a template that
- constructs a location on the currentTape when constructed
- tracks the location ownership in move assignment and move constructor in a `valid_` data member
- frees the location if the location has the ownership at destruction time

The template parameter is used to distinguish between location handling functions for `pdouble`
and `adouble` at compile-time. To this end, the template is restricted via the concept `adouble_or_pdouble`
since only these types are supported.

Note: `pdouble`s `free_loc` is not called since it apparently leads to errors. However, this is fine,
because future more performant StoreManager might never free locations and provide empty `free_loc`
functions.