Make the global buffers thread_local to enable basic openmp
The global tapeBuffer and tapeIdBuffer is now thread local, allowing together with the other changes the use of parallel taping and derivative computation where each thread has there own tapes!

In contrast to the previous openmp support, we dont have to compile adolc against openmp.

The files that are ocassionally used to store tape data contain from now on a thread id. This thread id is provide by the getThreadId function, which shares an atomic counter between all threads.

It was necessary to keep also the tapeInfos.stats data in the "initTapeInfos_keep" method.

Please Note: The notification in setCurrentTape is now removed.