Refactor fp_iteration; introduce functionalities for second order fixed-point computations
This PR introduces several modernizations like encapsulation of common methods to reduce code dublication or re-organize the code to have a clear meaning. There is now a FpMode enum which is used as template argument for fp_iteration to enable support for second order fixed-point computations. The theory of second order fp is based on the paper by Oliver sander and Andrea walther.