Split openmp test into separate cmake test
Thread sanitizer complained that memory used a some point in the parallel code for other tests eventhough those test not connected to the openmp test. There is still a warning from the sanitizer complaining about the destruction of the vector used in the omp region, but this is related to openmp and not adolc.